+++ packages/bluesky/lib/src/services/graph_service.dart
@@ -16,6 +16,7 @@ import 'entities/follows.dart';
 import 'entities/list_items.dart';
 import 'entities/lists.dart';
 import 'entities/mutes.dart';
+import 'entities/relationships.dart';
 import 'entities/suggested_follows.dart';
 import 'params/list_item_param.dart';
 import 'params/list_param.dart';
@@ -232,6 +233,20 @@ final class GraphService {
       await createBlockList(
         listUri: listUri,
         createdAt: createdAt,
+      );
+
+  /// https://atprotodart.com/docs/lexicons/app/bsky/graph/getRelationships
+  Future<core.XRPCResponse<Relationships>> getRelationships({
+    required String actor,
+    List<String>? others,
+  }) async =>
+      await _ctx.get(
+        ns.appBskyGraphGetRelationships,
+        parameters: {
+          'actor': actor,
+          'others': others,
+        },
+        to: Relationships.fromJson,
       );
 
   @Deprecated('Use .follow instead. Will be removed.')